Ship Milestone 1 firmware and fix GitHub 404s.
avr-ultra is now a scheduled fail-closed controller (close pulse on
reset, DS3231 wake, gated sensing, CRC EEPROM, six-layer safety)
instead of a Milestone 0 sleep skeleton.

Add the missing Makefile, AVR toolchain file, and wiring diagram so
those README links stop 404ing. make test runs host unit tests.

+++ MoistureController.ino
@@ -31,8 +31,8 @@
 volatile bool wakeUp = false;
 volatile int wdtCount = 0;
 const int HYSTERESIS = 10;  // ~1% offset for anti-chatter
-const int WATER_DURATION = 10000;  // 10 seconds
-const int MAX_WAIT_TIME = 60000;   // Safety timeout for hysteresis wait (60s)
+const unsigned long WATER_DURATION = 10000UL;  // 10 seconds
+const unsigned long MAX_WAIT_TIME = 60000UL;   // Safety timeout for hysteresis wait (60s)
 const long TARGET_SLEEP_MS = 300000;  // 5 min target, approximated with 8s WDT ticks
 
 void setup() {
